deleteImage

+++ classes/Project.php
@@ -21,7 +21,9 @@
             $div = explode('.', $file_name);
             $file_ext = strtolower(end($div));
             $unique_image = substr(md5(time()),0,10).'.'.$file_ext;
+            if (!empty($file_ext)){
             $upload_image = "upload/".$unique_image;
+            }else {$upload_image ="";}
 
             if (empty($name) || empty($description)){
                 $msg = "Поля должны быть заполнены";
@@ -34,7 +36,7 @@
 
                 $query = "INSERT INTO projects(name, description, image) VALUES ('$name', '$description', '$upload_image')";
 
-                $result = $this->db->insert($query);
+                $result = $this->db->query($query);
 
                 if ($result) {
                     $msg = "Редактирование успешно";
@@ -48,27 +50,29 @@
 
         public function allProject(){
             $query = "SELECT * FROM projects ORDER BY id";
-            $result = $this->db->select($query);
+            $result = $this->db->query($query);
             return $result;
         }
 
         public function getProjectById($id){
             $query = "SELECT * FROM projects WHERE id = '$id'";
-            $result =$this->db->select($query);
+            $result =$this->db->query($query);
             return $result;
         }
 
         public function deleteProject($extract_id){
             $query = "SELECT * FROM projects WHERE id IN($extract_id)";
-            $result=$this->db->select($query);
+            $result=$this->db->query($query);
             if($result) {
                 while ($row = $result->fetch()) {
                     $image =$row['image'];
+                    if (!empty($image)){
                     unlink($image);
+                    }
                 }
             }
             $del_query = "DELETE FROM projects WHERE id IN($extract_id)";
-            $del = $this->db->delete($del_query);
+            $del = $this->db->query($del_query);
         }
 
         public function updateProject($data, $file, $id){
@@ -82,7 +86,9 @@
             $div = explode('.', $file_name);
             $file_ext = strtolower(end($div));
             $unique_image = substr(md5(time()),0,10).'.'.$file_ext;
+            if (!empty($file_ext)){
             $upload_image = "upload/".$unique_image;
+            }else {$upload_image ="";}
 
             if (empty($name) || empty($description)){
                 $msg = "Поля должны быть заполнены";
@@ -94,11 +100,13 @@
                 }else {
 
                     $img_query = "SELECT * FROM projects WHERE id = '$id'";
-                    $img_res = $this->db->select($img_query);
+                    $img_res = $this->db->query($img_query);
                     if ($img_res) {
                         while ($row = $img_res->fetch()){
                             $image = $row['image'];
+                            if (!empty($image)){
                             unlink($image);
+                            }
                         }
                     }
                 
@@ -107,7 +115,7 @@
 
                     $query = "UPDATE projects SET name='$name', description='$description', image='$upload_image' WHERE id = '$id'";
 
-                    $result = $this->db->update($query);
+                    $result = $this->db->query($query);
 
                     if ($result) {
                        $msg = "Редактирование успешно";
@@ -120,7 +128,7 @@
             }else {
                 $query = "UPDATE projects SET name='$name', description='$description' WHERE id = '$id'";
 
-                $result = $this->db->update($query);
+                $result = $this->db->query($query);
 
                 if ($result) {
                     $msg = "Редактирование успешно";
@@ -131,6 +139,22 @@
                 }
             }
 
+        }
+
+        public function deleteImage($id){
+
+            $query = "SELECT * FROM projects WHERE id=$id";
+            $result=$this->db->query($query);
+            if($result) {
+                while ($row = $result->fetch()) {
+                    $image =$row['image'];
+                    if (!empty($image)){
+                    unlink($image);
+                    }
+                }
+            }
+            $del_query = "UPDATE projects SET image='' WHERE id='$id'";
+            $del = $this->db->query($del_query);
         }
 
     }
